# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API do Site da Donna Lupe",
        "title": "Donna Lupe Site API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
...
        "/admin/ingredients": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista o estoque de ingredientes cadastrados",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingredients"
                ],
                "summary": "Lista ingredientes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cadastra um novo ingrediente do estoque",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingredients"
                ],
                "summary": "Cadastra ingrediente",
                "parameters": [
                    {
                        "description": "Dados do ingrediente",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateIngredientRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/ingredients/{name}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza estoque, unidade ou valor de um ingrediente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingredients"
                ],
                "summary": "Atualiza ingrediente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome do ingrediente",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados para atualização",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateIngredientRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove um ingrediente e seus vínculos em product_ingredients",
                "tags": [
                    "ingredients"
                ],
                "summary": "Remove ingrediente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome do ingrediente",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/landing": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "landing"
                ],
                "summary": "Lista conteúdos da landing page",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "landing"
                ],
                "summary": "Cria conteúdo da landing page",
                "parameters": [
                    {
                        "description": "Conteúdo da landing",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LandingContentInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/landing/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "landing"
                ],
                "summary": "Atualiza conteúdo da landing page",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do conteúdo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Conteúdo atualizado",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LandingContentInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "landing"
                ],
                "summary": "Remove conteúdo da landing page",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do conteúdo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateIngredientRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "stock": {
                    "type": "number",
                    "minimum": 0
                },
                "unit": {
                    "type": "string"
                },
                "value_reais": {
                    "type": "number",
                    "minimum": 0
                }
            }
        },
...
        "models.LandingContentInput": {
            "type": "object",
            "required": [
                "name",
                "secao",
                "titulo"
            ],
            "properties": {
                "botaoLink": {
                    "type": "string"
                },
                "botaoTexto": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string"
                },
                "imagem": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "secao": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "subtitulo": {
                    "type": "string"
                },
                "titulo": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateIngredientRequest": {
            "type": "object",
            "properties": {
                "stock": {
                    "type": "number",
                    "minimum": 0
                },
                "unit": {
                    "type": "string"
                },
                "value_reais": {
                    "type": "number",
                    "minimum": 0
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}